# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Disaster Preparedness and Response Education System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List all learning modules
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LearningModule"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create a learning module (teacher/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningModuleInput"
      responses:
        "201":
          description: Created module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningModule"
  /modules/{id}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get a learning module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningModule"
        "404":
          description: Not found
    patch:
      operationId: updateModule
      tags: [modules]
      summary: Update a learning module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningModuleUpdate"
      responses:
        "200":
          description: Updated module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningModule"
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete a learning module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List active alerts
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create an alert (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete an alert (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incidents (admin sees all, others see own)
      responses:
        "200":
          description: Incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Report a new incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}/status:
    patch:
      operationId: updateIncidentStatus
      tags: [incidents]
      summary: Update incident status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentStatusUpdate"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /emergency-contacts:
    get:
      operationId: listEmergencyContacts
      tags: [emergency-contacts]
      summary: List emergency contacts
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Emergency contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyContact"
    post:
      operationId: createEmergencyContact
      tags: [emergency-contacts]
      summary: Create emergency contact (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyContactInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContact"
...
components:
  schemas:
...
    LearningModule:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [earthquake, flood, fire, cyclone, pandemic, landslide, heatwave]
        contentUrl:
          type: ["string", "null"]
        contentType:
          type: string
          enum: [video, pdf, article, guide]
        createdBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - contentType
        - createdAt
    LearningModuleInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [earthquake, flood, fire, cyclone, pandemic, landslide, heatwave]
        contentUrl:
          type: string
        contentType:
          type: string
          enum: [video, pdf, article, guide]
      required:
        - title
        - description
        - category
        - contentType
    LearningModuleUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [earthquake, flood, fire, cyclone, pandemic, landslide, heatwave]
        contentUrl:
          type: string
        contentType:
          type: string
          enum: [video, pdf, article, guide]
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        alertType:
          type: string
          enum: [weather, earthquake, fire, flood, evacuation, general]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - message
        - alertType
        - isActive
        - createdAt
    AlertInput:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        alertType:
          type: string
          enum: [weather, earthquake, fire, flood, evacuation, general]
      required:
        - title
        - message
        - alertType
    Incident:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        title:
          type: string
        description:
          type: string
        location:
          type: string
        incidentType:
          type: string
          enum: [fire, flood, building_collapse, medical_emergency, earthquake, other]
        status:
          type: string
          enum: [pending, investigating, resolved, closed]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - title
        - description
        - location
        - incidentType
        - status
        - createdAt
    IncidentInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        incidentType:
          type: string
          enum: [fire, flood, building_collapse, medical_emergency, earthquake, other]
        imageUrl:
          type: string
      required:
        - title
        - description
        - location
        - incidentType
    IncidentStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, investigating, resolved, closed]
      required:
        - status
    EmergencyContact:
      type: object
      properties:
        id:
          type: integer
        department:
          type: string
        contactNumber:
          type: string
        city:
          type: string
        contactType:
          type: string
          enum: [police, fire, medical, disaster_management, utility, other]
      required:
        - id
        - department
        - contactNumber
        - city
        - contactType
    EmergencyContactInput:
      type: object
      properties:
        department:
          type: string
        contactNumber:
          type: string
        city:
          type: string
        contactType:
          type: string
          enum: [police, fire, medical, disaster_management, utility, other]
      required:
        - department
        - contactNumber
        - city
        - contactType
...