# lib/api-zod/src/generated/types/emergencyContactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Disaster Preparedness and Response Education System API
 * OpenAPI spec version: 0.1.0
 */
import type { EmergencyContactInputContactType } from "./emergencyContactInputContactType";

export interface EmergencyContactInput {
  department: string;
  contactNumber: string;
  city: string;
  contactType: EmergencyContactInputContactType;
}