# lib/api-zod/src/generated/types/learningModuleInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Disaster Preparedness and Response Education System API
 * OpenAPI spec version: 0.1.0
 */
import type { LearningModuleInputCategory } from "./learningModuleInputCategory";
import type { LearningModuleInputContentType } from "./learningModuleInputContentType";

export interface LearningModuleInput {
  title: string;
  description: string;
  category: LearningModuleInputCategory;
  contentUrl?: string;
  contentType: LearningModuleInputContentType;
}